requirements.txt
streamlit~=1.15.2
numpy~=1.19.5
pandas~=1.3.1
tensorflow-gpu~=2.5.0
keras~=2.4.3
opencv-python~=4.4.0.44
pillow~=8.2.0
openpyxl~=3.0.10
tqdm~=4.51.0
matplotlib~=3.4.2
torch~=1.9.0+cu111
torchvision~=0.10.0+cu111
dlib~=19.17.99
flask~=1.1.2
requests~=2.25.1
setuptools~=52.0.0